If a dc:accessRights tag exists in the feed, assign visibility
If the visibility is not specified, we assume public. If it is specified with an incorrect value, we assume private. The only accepted values at this time are private or public. The result is assigned as the wnpa_item_visibility taxonomy to the feed item as it is pulled in.

Closes #18