Restructure directory to be a proper WordPress plugin
Until now, the project has been a combination of the wnpa.wsu.edu
website and the open source WNPA Syndication plugin. This made
sense initially, when we were deploying WNPA to a single server
and weren't sure what the project goals were. Now that things are
a bit more focused, we can setup a different repository for the
theme and maintain this as a WordPress plugin.